Generics.Kind

Description

Main module of kind-generics. Please refer to the README file for documentation on how to use this package.

Generic representation types

data ((f :: k -> Type) :+: (g :: k -> Type)) (p :: k) infixr 5

Sums: encode choice between constructors

Constructors

L1 (f p)

R1 (g p)

data ((f :: k -> Type) :*: (g :: k -> Type)) (p :: k) infixr 6

Products: encode multiple arguments to constructors

Constructors

(f p) :*: (g p) infixr 6

data V1 (p :: k)

Void: used for datatypes without constructors

data U1 (p :: k)

Unit: used for constructors without arguments

Constructors

U1

newtype M1 i (c :: Meta) (f :: k -> Type) (p :: k)

Meta-information (constructor names, etc.)

Constructors

M1

Fields

unM1 :: f p

newtype Field (t :: Atom d Type) (x :: LoT d) where

Fields: used to represent each of the (visible) arguments to a constructor. Replaces the K1 type from Generics. The type of the field is represented by an Atom from Atom.

instance GenericK [] (a :&&: LoT0) where
  type RepK [] = Field Var0 :*: Field ([] :$: Var0)

Constructors

Field

Fields

:: forall {d} (t :: Atom d Type) (x :: LoT d). { unField :: Interpret t x

} -> Field t x

data ((c :: Atom d Constraint) :=>: (f :: LoT d -> Type)) (x :: LoT d) where

Constraints: used to represent constraints in a constructor. Replaces the (:=>:) type from GHC.Generics.Extra.

data Showable a = Show a => a -> X a

instance GenericK Showable (a :&&: LoT0) where
  type RepK Showable = (Show :$: a) :=>: (Field Var0)

Constructors

SuchThat :: forall {d} (c :: Atom d Constraint) (x :: LoT d) (f :: LoT d -> Type). Interpret c x => f x -> (c :=>: f) x

data Exists k (f :: LoT (k -> d) -> Type) (x :: LoT d) where

Existentials: a representation of the form E f describes a constructor whose inner type is represented by f, and where the type variable at index 0, Var0, is existentially quantified.

data E where
 E :: t -> Exists

instance GenericK E LoT0 where
  type RepK E = Exists Type (Field Var0)

Constructors

Exists

Fields

:: forall k (t :: k) d (f :: LoT (k -> d) -> Type) (x :: LoT d). { unExists :: f (t ':&&: x)

} -> Exists k f x

Atoms for Field

data Atom d (k :: TYPE r) where

Shape of a type, possibly with type variables.

>>> -- the type [a] for unknown a
>>> :kind Kon [] :@: Var0
Kon [] :@: Var0 :: Atom (Type -> xs) Type

Representation of type families

Expand

Type families are represented using first-class-families.

For example, the type-level n + m :: Nat-- may expand to the following--

n + m         -- using (+) from GHC.TypeNats
~
Eval (n + m)  -- using Eval from Fcf.Core and (+) from Fcf.Data.Nat

which may be encoded as the following Atom (using Var0 for n and Var1 for m):

Eval ((Kon (+) :@: Var0) :@: Var1)  -- Eval as Atom's constructor
  :: Atom (Nat -> Nat -> Type) Nat

kind-generics uses a different, more systematic encoding of type families for GenericK instances; see fcf-family for more details. For example, n + m is instead expanded to the following:

n + m
~
Eval (NDFamily (MkName "base" "GHC.TypeNats" "+") P0 '(n, '(m, '())))

which gets encoded as the following Atom:

Eval (Kon (NDFamily (MkName "base" "GHC.TypeNats" "+") P0)
        :@: ((Kon '(,) :@: Var0) :@: ((Kon '(,) :@: Var1) :@: Kon '())))
  :: Atom (Nat -> Nat -> Type) Nat

Constructors

Var :: forall d k1. TyVar d k1 -> Atom d k1
Represents a type variable.

Kon :: forall k1 d. k1 -> Atom d k1
Represents a constant type, like Int.

(:@:) :: forall d k1 k2. Atom d (k1 -> k2) -> Atom d k1 -> Atom d k2
Represents type application.

(:&:) :: forall d. Atom d Constraint -> Atom d Constraint -> Atom d Constraint infixr 5
Represents the conjunction of two constraints.

ForAll :: forall d1 d. Atom (d1 -> d) Type -> Atom d Type
Represents universal quantification.

(:=>>:) :: forall d. Atom d Constraint -> Atom d Type -> Atom d Type infixr 5
Represents constraint requirement, the "thick arrow" =>.

Eval :: forall d k1. Atom d (Exp k1) -> Atom d k1
Represents a type family application.

data TyVar d (k :: TYPE r) where

Well-scoped de Bruijn representation of type variables. TyVar d represents all the possible type variables which can refer to the holes in kind d.

We recommend using the aliases Var0, Var1, ... instead of the constructors, for further clarity.

Constructors

VZ :: forall x xs. TyVar (x -> xs) x
First hole in d.

VS :: forall xs k1 x. TyVar xs k1 -> TyVar (x -> xs) k1
Successor hole, increases the hole reference by 1.

type (:$:) (f :: k1 -> k2) (x :: Atom d k1) = ('Kon f :: Atom d (k1 -> k2)) ':@: x

Represents an applied constructor. Instead of Kon [] :: Var0$ you can write @[] :$: Var0$.

type (:~:) (a :: Atom d k1) (b :: Atom d k1) = (('Kon ((~) :: k1 -> k1 -> Constraint) :: Atom d (k1 -> k1 -> Constraint)) ':@: a) ':@: b

Represents (homogeneous) type equality.

type (:~~:) (a :: Atom d k1) (b :: Atom d k2) = (('Kon ((~~) :: k1 -> k2 -> Constraint) :: Atom d (k1 -> k2 -> Constraint)) ':@: a) ':@: b

Represents heterogeneous type equality.

type Var0 = 'Var ('VZ :: TyVar (k -> xs) k)

type Var1 = 'Var ('VS ('VZ :: TyVar (k -> xs) k) :: TyVar (x -> k -> xs) k)

type Var2 = 'Var ('VS ('VS ('VZ :: TyVar (k -> xs) k) :: TyVar (x1 -> k -> xs) k) :: TyVar (x -> x1 -> k -> xs) k)

type Var3 = 'Var ('VS ('VS ('VS ('VZ :: TyVar (k -> xs) k) :: TyVar (x2 -> k -> xs) k) :: TyVar (x1 -> x2 -> k -> xs) k) :: TyVar (x -> x1 -> x2 -> k -> xs) k)

type Var4 = 'Var ('VS ('VS ('VS ('VS ('VZ :: TyVar (k -> xs) k) :: TyVar (x3 -> k -> xs) k) :: TyVar (x2 -> x3 -> k -> xs) k) :: TyVar (x1 -> x2 -> x3 -> k -> xs) k) :: TyVar (x -> x1 -> x2 -> x3 -> k -> xs) k)

type Var5 = 'Var ('VS ('VS ('VS ('VS ('VS ('VZ :: TyVar (k -> xs) k) :: TyVar (x4 -> k -> xs) k) :: TyVar (x3 -> x4 -> k -> xs) k) :: TyVar (x2 -> x3 -> x4 -> k -> xs) k) :: TyVar (x1 -> x2 -> x3 -> x4 -> k -> xs) k) :: TyVar (x -> x1 -> x2 -> x3 -> x4 -> k -> xs) k)

type Var6 = 'Var ('VS ('VS ('VS ('VS ('VS ('VS ('VZ :: TyVar (k -> xs) k) :: TyVar (x5 -> k -> xs) k) :: TyVar (x4 -> x5 -> k -> xs) k) :: TyVar (x3 -> x4 -> x5 -> k -> xs) k) :: TyVar (x2 -> x3 -> x4 -> x5 -> k -> xs) k) :: TyVar (x1 -> x2 -> x3 -> x4 -> x5 -> k -> xs) k) :: TyVar (x -> x1 -> x2 -> x3 -> x4 -> x5 -> k -> xs) k)

type Var7 = 'Var ('VS ('VS ('VS ('VS ('VS ('VS ('VS ('VZ :: TyVar (k -> xs) k) :: TyVar (x6 -> k -> xs) k) :: TyVar (x5 -> x6 -> k -> xs) k) :: TyVar (x4 -> x5 -> x6 -> k -> xs) k) :: TyVar (x3 -> x4 -> x5 -> x6 -> k -> xs) k) :: TyVar (x2 -> x3 -> x4 -> x5 -> x6 -> k -> xs) k) :: TyVar (x1 -> x2 -> x3 -> x4 -> x5 -> x6 -> k -> xs) k) :: TyVar (x -> x1 -> x2 -> x3 -> x4 -> x5 -> x6 -> k -> xs) k)

type Var8 = 'Var ('VS ('VS ('VS ('VS ('VS ('VS ('VS ('VS ('VZ :: TyVar (k -> xs) k) :: TyVar (x7 -> k -> xs) k) :: TyVar (x6 -> x7 -> k -> xs) k) :: TyVar (x5 -> x6 -> x7 -> k -> xs) k) :: TyVar (x4 -> x5 -> x6 -> x7 -> k -> xs) k) :: TyVar (x3 -> x4 -> x5 -> x6 -> x7 -> k -> xs) k) :: TyVar (x2 -> x3 -> x4 -> x5 -> x6 -> x7 -> k -> xs) k) :: TyVar (x1 -> x2 -> x3 -> x4 -> x5 -> x6 -> x7 -> k -> xs) k) :: TyVar (x -> x1 -> x2 -> x3 -> x4 -> x5 -> x6 -> x7 -> k -> xs) k)

type Var9 = 'Var ('VS ('VS ('VS ('VS ('VS ('VS ('VS ('VS ('VS ('VZ :: TyVar (k -> xs) k) :: TyVar (x8 -> k -> xs) k) :: TyVar (x7 -> x8 -> k -> xs) k) :: TyVar (x6 -> x7 -> x8 -> k -> xs) k) :: TyVar (x5 -> x6 -> x7 -> x8 -> k -> xs) k) :: TyVar (x4 -> x5 -> x6 -> x7 -> x8 -> k -> xs) k) :: TyVar (x3 -> x4 -> x5 -> x6 -> x7 -> x8 -> k -> xs) k) :: TyVar (x2 -> x3 -> x4 -> x5 -> x6 -> x7 -> x8 -> k -> xs) k) :: TyVar (x1 -> x2 -> x3 -> x4 -> x5 -> x6 -> x7 -> x8 -> k -> xs) k) :: TyVar (x -> x1 -> x2 -> x3 -> x4 -> x5 -> x6 -> x7 -> x8 -> k -> xs) k)

Lists of types

data LoT k where

LoT k represents a list of types which can be applied to a data type of kind k.

Constructors

LoT0 :: LoT Type
Empty list of types.

(:&&:) :: forall k1 ks. k1 -> LoT ks -> LoT (k1 -> ks) infixr 5
Cons a type with a list of types.

type family (f :: k) :@@: (tys :: LoT k) where ...

Apply a list of types to a type constructor.

>>> :kind! Either :@@: (Int :&&: Bool :&&: LoT0)
Either :@@: (Int :&&: Bool :&&: LoT0) :: Type
= Either Int Bool

Equations

(f :: Type) :@@: (_1 :: LoT Type) = f

(f :: k -> k') :@@: (as :: LoT (k -> k')) = f (HeadLoT as) :@@: TailLoT as

type LoT1 (a :: k) = a ':&&: 'LoT0

List of types with a single element.

type LoT2 (a :: k) (b :: k1) = a ':&&: (b ':&&: 'LoT0)

List of types with two elements.

data TyEnv where

A type constructor and a list of types that can be applied to it.

Constructors

TyEnv :: forall k. k -> LoT k -> TyEnv

Generic type classes

class GenericK (f :: k) where

Representable types of any kind. Examples:

instance GenericK Int
instance GenericK []
instance GenericK Either
instance GenericK (Either a)
instance GenericK (Either a b)

Minimal complete definition

Nothing

Associated Types

type RepK (f :: k) :: LoT k -> Type

Methods

fromK :: forall (x :: LoT k). (f :@@: x) -> RepK f x

Convert the data type to its representation.

default fromK :: forall (x :: LoT k). (Generic (f :@@: x), Conv (Rep (f :@@: x)) (RepK f) x) => (f :@@: x) -> RepK f x

toK :: forall (x :: LoT k). RepK f x -> f :@@: x

Convert from a representation to its corresponding data type.

default toK :: forall (x :: LoT k). (Generic (f :@@: x), Conv (Rep (f :@@: x)) (RepK f) x) => RepK f x -> f :@@: x

type GenericF t (f :: k) (x :: LoT k) = (GenericK f, x ~ SplitF t f, t ~ (f :@@: x))

fromF :: forall {k} (f :: k) t (x :: LoT k). GenericF t f x => t -> RepK f x

toF :: forall {k} (f :: k) t (x :: LoT k). GenericF t f x => RepK f x -> t

type GenericN (n :: Nat) t (f :: k) (x :: LoT k) = (GenericK f, 'TyEnv f x ~ SplitN n t, t ~ (f :@@: x))

fromN :: forall {k} (n :: Nat) t (f :: k) (x :: LoT k). GenericN n t f x => t -> RepK f x

toN :: forall {k} (n :: Nat) t (f :: k) (x :: LoT k). GenericN n t f x => RepK f x -> t

Getting more instances almost for free

fromRepK :: forall {t} {k} (f :: t -> k) (x :: t) (xs :: LoT k). (GenericK f, SubstRep' (RepK f) x xs) => (f x :@@: xs) -> SubstRep (RepK f) x xs

toRepK :: forall {t} {k} (f :: t -> k) (x :: t) (xs :: LoT k). (GenericK f, SubstRep' (RepK f) x xs) => SubstRep (RepK f) x xs -> f x :@@: xs

type family SubstRep (f :: LoT (t -> k) -> Type) (x :: t) :: LoT k -> Type

class SubstRep' (f :: LoT (t -> k) -> Type) (x :: t) (xs :: LoT k)

Minimal complete definition

substRep, unsubstRep

type family SubstAtom (f :: Atom (t -> k) d) (x :: t) :: Atom k d where ...

Equations

SubstAtom ('Var ('VZ :: TyVar (k1 -> k) k1) :: Atom (k1 -> k) k1) (x :: k1) = 'Kon x :: Atom k k1

SubstAtom ('Var ('VS v :: TyVar (t -> d) k1) :: Atom (t -> d) k1) (x :: t) = 'Var v

SubstAtom ('Kon t2 :: Atom (t1 -> k) k1) (x :: t1) = 'Kon t2 :: Atom k k1

SubstAtom (t1 ':@: t2 :: Atom (t -> k) k2) (x :: t) = SubstAtom t1 x ':@: SubstAtom t2 x

SubstAtom (t1 ':&: t2 :: Atom (t -> k) Constraint) (x :: t) = SubstAtom t1 x ':&: SubstAtom t2 x

Bridging with GHC.Generics

class Conv (gg :: Type -> Type) (kg :: LoT d -> Type) (tys :: LoT d) where

Bridges a representation of a data type using the combinators in GHC.Generics with a representation using this module. You are never expected to manipulate this type class directly, it is part of the deriving mechanism for GenericK.

Methods

toGhcGenerics :: kg tys -> gg a

toKindGenerics :: gg a -> kg tys

Re-exported from Atom

Interpretation of atoms

type family Interpret (t :: Atom d k) (tys :: LoT d) :: k where ...

Replaces the holes in the Atom t by the elements of the list of types tys. The amount and kind of types in tys must match statically those required by the Atom.

>>> :kind! Interpret ([] :$: Var0) (LoT1 Int)
Interpret ([] :$: Var0) (LoT1 Int) :: Type
= [Int]

Equations

Interpret ('Var v :: Atom d k) (tys :: LoT d) = InterpretVar v tys

Interpret ('Kon t :: Atom d k) (tys :: LoT d) = t

Interpret (f ':@: x :: Atom d k2) (tys :: LoT d) = Interpret f tys (Interpret x tys)

Interpret (c ':&: d2 :: Atom d1 Constraint) (tys :: LoT d1) = (Interpret c tys, Interpret d2 tys)

Interpret ('ForAll f :: Atom d Type) (tys :: LoT d) = ForAllI f tys

Interpret (c ':=>>: f :: Atom d Type) (tys :: LoT d) = SuchThatI c f tys

Interpret ('Eval f :: Atom d a) (tys :: LoT d) = Eval (Interpret f tys)

type family InterpretVar (t :: TyVar d k) (tys :: LoT d) :: k where ...

Obtains the type in the list tys referenced by the type variable t.

>>> :kind! Interpret Var0 (LoT2 Int Bool)
Interpret Var0 (LoT2 Int Bool) :: Type
= Int
>>> :kind! Interpret Var1 (LoT2 Int Bool)
Interpret Var1 (LoT2 Int Bool) :: Type
= Bool

Equations

InterpretVar ('VZ :: TyVar (k -> k') k) (tys :: LoT (k -> k')) = HeadLoT tys

InterpretVar ('VS v :: TyVar (k2 -> k') k1) (tys :: LoT (k2 -> k')) = InterpretVar v (TailLoT tys)

type family Satisfies (cs :: [Atom d Constraint]) (tys :: LoT d) where ...

Interprets a list of Atom representing constraints into the actual constraints. This is a specialization of Interpret for the case of constraints.

>>> :kind! Satisfies '[Eq :$: Var0, Show :$: Var0] (LoT1 Int)
Satisfies '[Eq :$: Var0, Show :$: Var0] (LoT1 Int) :: Constraint
= (Eq Int, (Show Int, () :: Constraint))

Equations

Satisfies ('[] :: [Atom d Constraint]) (tys :: LoT d) = ()

Satisfies (c ': cs :: [Atom d Constraint]) (tys :: LoT d) = (Interpret c tys, Satisfies cs tys)

type family ContainsTyVar (v :: TyVar d k) (t :: Atom d p) :: Bool where ...

Determines whether a given type variable v is used within an Atom t. If not, we know that the atom is constant with respect to that variable.

Equations

ContainsTyVar (v :: TyVar d p) ('Var v :: Atom d p) = 'True

ContainsTyVar (v :: TyVar d k) ('Var w :: Atom d p) = 'False

ContainsTyVar (v :: TyVar d k) ('Kon t :: Atom d p) = 'False

ContainsTyVar (v :: TyVar d k) (f ':@: x :: Atom d p2) = Or (ContainsTyVar v f) (ContainsTyVar v x)

ContainsTyVar (v :: TyVar d k) (x ':&: y :: Atom d Constraint) = Or (ContainsTyVar v x) (ContainsTyVar v y)

ContainsTyVar (v :: TyVar d k) (c ':=>>: f :: Atom d Type) = Or (ContainsTyVar v c) (ContainsTyVar v f)

ContainsTyVar (v :: TyVar xs k) ('ForAll f :: Atom xs Type) = ContainsTyVar ('VS v :: TyVar (x -> xs) k) f

Auxiliary data types for interpretation

newtype ForAllI (f :: Atom (d1 -> d) Type) (tys :: LoT d) where

Auxiliary type for interpretation of the ForAll atom. Required because a type family like Interpret cannot return a polymorphic type.

Constructors

ForAllI :: forall {d1} {d} (f :: Atom (d1 -> d) Type) (tys :: LoT d). (forall (t :: d1). Interpret f (t ':&&: tys)) -> ForAllI f tys

newtype SuchThatI (c :: Atom d Constraint) (f :: Atom d Type) (tys :: LoT d) where

Auxiliary type for interpretation of the (:=>>:) atom. Required because a type family like Interpret cannot return a type with constraints.

Constructors

SuchThatI

Fields

:: forall {d} (c :: Atom d Constraint) (tys :: LoT d) (f :: Atom d Type). { unSuchThatI :: Interpret c tys => Interpret f tys

} -> SuchThatI c f tys

newtype WrappedI (f :: Atom d Type) (tys :: LoT d)

Records a value of type f applied to the list tys.

>>> :t WrapI [1] :: WrappedI ([] :$: Var0) (LoT1 Int)
WrapI [1] :: WrappedI ([] :$: Var0) (LoT1 Int)
  :: WrappedI ([] :$: Var0) (LoT1 Int)

Constructors

WrapI

Fields

unwrapI :: Interpret f tys

toWrappedI :: forall {d1} {ks} (f :: Atom (d1 -> ks) Type) (tys :: LoT ks) (t :: d1). ForAllI f tys -> WrappedI f (t ':&&: tys)

fromWrappedI :: forall {d1} {d} (f :: Atom (d1 -> d) Type) (tys :: LoT d). (forall (t :: d1). WrappedI f (t ':&&: tys)) -> ForAllI f tys